\name{weightedCL-package}
\alias{weightedCL-package}
\docType{package}
\title{Efficient and feasible inference for high-dimensional normal copula regression models}
\description{
The weighted composite likelihood estimating equations for high-dimensional normal copula regression models in Nikoloulopoulos (2022). 
}
\details{
This package contains \R functions  to estimate high-dimensional MVN copula regression models  with the WCL estimating equations in Nikoloulopoulos (2022). It provides ARMA\eqn{(p, q)} correlation structures and binary, ordinal,  Poisson, and  negative binomial (both NB1 and NB2 parametrizations) regressions.
}

\author{
Aristidis K. Nikoloulopoulos.
}
\references{

Nikoloulopoulos, A.K. (2022)
Efficient and feasible inference for high-dimensional normal copula regression models.
 \emph{Arxiv e-prints}, <arXiv:2203.04619>. \url{https://arxiv.org/abs/2203.04619}.

}
\keyword{package}
