\name{weightedCL-package}
\alias{weightedCL-package}
\docType{package}
\title{Efficient and feasible inference for high-dimensional normal copula regression models}
\description{
The weighted composite likelihood estimating equations for high-dimensional normal copula regression models in Nikoloulopoulos (2023). 
}
\details{
This package contains \R functions  to estimate high-dimensional MVN copula regression models  with the WCL estimating equations in Nikoloulopoulos (2022). It provides ARMA\eqn{(p, q)} correlation structures and binary, ordinal,  Poisson, and  negative binomial (both NB1 and NB2 parametrizations) regressions.
}

\author{
Aristidis K. Nikoloulopoulos.
}
\references{

Nikoloulopoulos, A.K. (2023) Efficient and feasible inference for high-dimensional normal copula regression models. \emph{Computational Statistics and  Data Analysis}, \bold{179}, 107654. \doi{10.1016/j.csda.2022.107654}.

}
\keyword{package}
